Diagnostic float: Split event handling flow
As different events actually need to trigger different actions, it is
easier to follow the logic when each event has its own sequence of
actions.

It also reduces the amount of work, as only |CursorMoved| can cause the
float to appear.

Restore |CursorHold| functionality disabled by

```gitcommit
commit 10fac72582a7f95ec3a0bf6715aaf0a7eb997bec
Author: suguruwataru <23465321+suguruwataru@users.noreply.github.com>
Date:   Thu Jan 1 19:09:58 2026 +0800

    Stop `CursorHold` events from updating diagnostic floating window (#1624)

    Closes #1620

    Co-authored-by: mattn <mattn.jp@gmail.com>
```

It is still nice to have an option of hiding the float when cursor did
not move long enough.  A subsequent commit adds a configuration option
to control this behavior.